requirement2.txt
streamlit==1.33.0
pandas==2.1.4
numpy==1.26.4
plotly==5.22.0
scipy==1.12.0
gcsfs==2024.2.0
google-cloud-storage==2.16.0
python-dotenv==1.0.1